+++ src/containers/Models.js
@@ -16,25 +16,7 @@ const Models = () => {
   const [errors, setErrors] = useState('');
   const [message, setMessage] = useState('');
 
-  const onSubmit = (data) => {
-    axios.post(`${urlBase}/api/v1/cars`, {
-      car: {
-        color: data.color,
-        engine: data.engine,
-        year: data.year,
-        title: data.title,
-        price: data.price,
-      },
-    }, { withCredentials: true })
-      .then((response) => {
-        if (response.data.status === 'created') {
-          dispatch(createCar(response.data));
-          history.push('/models');
-          setMessage('Created successfully close the form');
-        }
-        setErrors(response.data.errors);
-      });
-  };
+  
   useEffect(() => {
     dispatch(fetchCars());
   }, []);
@@ -43,81 +25,7 @@ const Models = () => {
       {
         userLogin.logged_in ? [
           userLogin.details.admin ? [
-            <div>
-              <Link to="/cars">see cars</Link>
-              <div className="container mt-5">
-                <button
-                  type="button"
-                  className="btn btn-primary"
-                  data-bs-toggle="modal"
-                  data-bs-target="#myModal"
-                >
-                  {' '}
-                  Add new car
-                </button>
-                <div className="modal" id="myModal">
-                  <div className="modal-dialog">
-                    <div className="modal-content">
-                      {message && (
-                      <div className="modal-header col-12">
-                        <div className="alert alert-success col-12" role="alert">{message}</div>
-                        <button type="button" className="btn-close alert-success close-button" data-bs-dismiss="modal" aria-label="close" />
-                      </div>
-                      )}
-                      <div className="modal-body">
-                        <form className="form" onSubmit={handleSubmit(onSubmit)}>
-                          {errors && <div className="alert alert-danger col-12" role="alert">{ errors.map((error) => <li className="text-start">{error}</li>) }</div>}
-                          <div className="form-floating mb-2 col-10">
-                            <input
-                              type="text"
-                              className="form-control"
-                              name="color"
-                              {...register('color', { required: true })}
-                              id="floatingInput"
-                              placeholder="Car color"
-                            />
-                            <label htmlFor="floatingInput">Color</label>
-                          </div>
-                          <div className="form-floating mb-2 col-10">
-                            <input type="text" name="engine" {...register('engine', { required: true })} className="form-control" id="floatingInputEmail" placeholder="Enter engine type" />
-                            <label htmlFor="floatingInputEmail">Engine</label>
-                          </div>
-                          <div className="form-floating mb-2 col-10">
-                            <input type="text" name="year" {...register('year', { required: true })} className="form-control" id="floatingPassword" placeholder="Enter Year" />
-                            <label htmlFor="floatingPassword">Year</label>
-                          </div>
-                          <div className="form-floating mb-2 col-10">
-                            <input
-                              type="text"
-                              name="title"
-                              {...register('title', { required: true })}
-                              className="form-control"
-                              id="floatingInputtitle"
-                              placeholder="Enter the Model"
-                            />
-                            <label htmlFor="floatingInputtitle">Model</label>
-                          </div>
-                          <div className="form-floating mb-2 col-10">
-                            <input
-                              type="text"
-                              name="price"
-                              {...register('price', { required: true })}
-                              className="form-control"
-                              id="floatingInputprice"
-                              placeholder="Enter the price"
-                            />
-                            <label htmlFor="floatingInputImage">Price</label>
-                          </div>
-                          <div className="form-floating mb-3 col-10">
-                            <button type="submit" className="btn btn-primary col-10">Add Car</button>
-                          </div>
-                        </form>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>,
+            
           ] : [
             <div>
               <CarList />
