+++ src/containers/NameList.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const NameList = ({ car }) => (
-  <div>
+  <div key={car.id}>
     <p>{car.color}</p>
     <p>{car.engine}</p>
   </div>
@@ -10,6 +10,7 @@ const NameList = ({ car }) => (
 
 NameList.propTypes = {
   car: PropTypes.shape({
+    id: PropTypes.number.isRequired,
     color: PropTypes.string.isRequired,
     engine: PropTypes.string.isRequired,
   }).isRequired,
